requirements.txt
matplotlib
nltk
notion-client
numpy
openai
python-dotenv
redis
tqdm
